refactor: unify OAuthTokenResponseError + collapse two asset files
Two bounded P2 collapses from the post-refactor audit:

1. OAuthTokenResponseError -> defineErrors

The only hand-rolled Error class in @epicenter/auth becomes a
defineErrors-style variant set, and parseOAuthTokenGrant returns
Result<OAuthTokenGrant, OAuthTokenResponseError> instead of throwing.
Call sites switch from `instanceof` + `.issue` to standard `error.name`
narrowing, matching every other error in the package.

  oauth-token-response.ts        class + throw  ->  defineErrors + Result
  oauth-launchers/index.ts       instanceof + switch on .issue
                                   ->  switch on .name
  node/oob-launcher.ts           try/catch  ->  destructure Result
  create-oauth-app-auth.ts       implicit throw  ->  destructure + rethrow
                                                    with explicit message

2. Merge asset-routes.ts into routes/assets.ts

The split between asset-routes.ts (top level, 325 lines) and
routes/assets.ts (40-line wrapper that just composes the two factories)
existed for a pre-collapse personal-vs-team URL story that no longer
applies. After the Owner partition collapse, both factories mount under
the same `/owners/:ownerId/assets` prefix, the mount-order rule is one
line, and there is no reason to spread them across two files in two
directories. Both factories are now file-private inside routes/assets.ts.

Deferred (taste calls, not in this commit):
- AuthState collapse (signed-in/reauth-required shared payload + bool):
  changes the discriminated-union shape every UI consumer relies on;
  deserves its own PR with a migration story.
- Single-load orchestration in machine-auth: two 1KB file reads is
  microseconds; the cleanup needs an API surface change to
  createMachineAuthClient that is not worth it.
- Inline derive-workspace-keyring.ts: 25 lines with two callers in
  sibling files; inlining duplicates or adds an import edge. Wash.
